Move deepEq implementation entirely to Haskell code.
Previously, the implementation was split between two functions in
two different worlds: a shell in codeworld-base, and a JavaScript
implementation at runtime.  This change moves the JavaScript into
codeworld-base alongside the FFI trappings, using multiline strings
to embed the implementation.

It's not pretty, and I'd like to work on further moving JavaScript
into Haskell there.  But it's far, far better than the previous
situation.