Commence work on a stand-alone mode
This is using the blank-canvas package to get a local webserver that
controls a Canvas. This way, the drawing semantics match very well to
what is happening on the server. The user still has to open the web
browser, though.

At this stage, only drawingOf is supported, and drawing the CodeWorld
logo will crash.