Tally number and size of broadcasted events per game
and at the end of the game, add that to a total count. (This is
currently slightly easier than adding them as they go, as this way the
game thread does not need acesses to the global total stats MVar. This
can be changed, of course.)

As a side effect, serialize the broadcast message only once, instead
once per player. Duh. (Unless GHC did that for us, but I doubt it.)

Part of #390.