Remove Pings and Timestamps from the codeworld-game-api
As discussed in #378, we only really care about timestamps generated by
the client, so the server does not need to handle them. If the clients
want to use them, they need to be included in the event payload.

Similarly, the server does not need to know whether an event is a ping
or not. The clients can just put an appropriate message into the event
payload.