Implement client-side timestamp logic (#378)
I does not yet cause the clients to send pings around. Also, the driver
code needs to be scrutinized for confusion of timestamps (relative to
the local start of the game) and absolute time. Or a newtype is in
order.

The whole business of using Doubles here makes me uneasy. Switching to
Integer might be in order.