Make the game join UI typesafe
Previously, in the driver, there was one constructor (`Setup`) used for
all stages of the setup, even if during some states some of the fields
did not make any sense. This is obviously not nice.

This now introduces a bit of type-level programming to index the UI
states by the setup phase (main, connecting, waiting), and allow the
driver to use different constructors.