Fix #457 Allow Users to Open Help in New Tab
When in iframe, creates a link at the top of help to open in new
tab. Additional code ensures the pages remain "linked" so clicking
code samples still loads the samples in the original page.

Despite my best efforts, I could not get the coding tab to focus
when a code sample was clicked. This feature appears to be
purposefully disabled in most browsers.
(see http://stackoverflow.com/questions/2758608/)