Fix default style after font size changer.
The earlier patch to enable keys that change font size changed the
default font size in many cases, apparently by accident.  This update
leaves the default style unchanged, and uses getComputedStyle to
make changes relative to the default.

One unfortunate aspect is that user stylesheets may set the default
in any unit.  We adjust by increments of 1 in whatever unit the original
value was specified in.  Hopefully, that will be 'px' or 'pt' in most
cases, in which case this behavior works.  But if it's something weird
like % or inches, this is going to behave oddly.  I'm okay with that.