FRP interface reativeOf
This patch allows users to create interactions by defining a FRP
program, based on the reflex library. The dependency on reflex is
optional, and enabled with -freflex.

This is a result of some hacking with the reflex people at the Compose
Unconference.

A mode where even reactive collaborations are supported depends on some
fancy new feature in reflex, which would allow stats to be reset.