Allow users to inspect drawingOf
The drawingOf entry point will register a callback with
initDebugMode allowing users to inspect pictures. Toggling the
inspect button will allow users to click on an image and be shown the
corresponding "stack" of pictures generating the element clicked.

For instance if a user clicks a polygon that has been colored,
translated and rotated a box will appear indicating this, and the
locations in code where the translation, rotation and creation of
the picture took place will appear yellow.