Add Android Export Handler
- exportAndroid handler which initializes new Cordova project

Add a new handler, `exportAndroid`. This takes the source as a
parameter, and performs the same steps that the compile route performs,
and calls `compileIfNeeded`.

After compilation, the `initCordovaProject` function is invoked, which
merely creates a stub cordova project, by making a subprocess call to
the `cordova` executable. At this commit, it does not build the project.