Resume accepting "program" as a synonym for "main".
The previous version of this change was commit 06c68ee, and was
reverted because it broke layout and column numbers.  The clever
trick here is not to define "program" to be "main", but rather to
be "(main )", which is the same number of characters as "program",
establishes layout in the same place, etc.  This hack should be
obsolete if a fixed -main-is can be merged into some future GHC.

As before, the plan is to verify this is working as expected, then
update documentation, then add a warning for code still using
"main", and finally to wait for a long time before breaking
anything.  In fact, might as well wait for a proper fix in GHC,
which looks like it could be a few years off.