Change how Picture information is communicated
This changes how information about picture is communication between
the javascript and haskell parts of debug mode. When debug mode
is now activated the javascript will get an object representing
the current image. When communicating about a specific node, such
as which node is being selected, the node's id (the order it appears
in DFS) is used.

This does not change functionality, but will be useful in the future.
This commit has added a few bugs that still need to be worked out.