Fixes for the client-side error sanitizer.
It turns out the old sanitizer was using regex-compat, which links
to standard libraries.  Fixed to rely only on regex-tdfa.  Also fixed
a few patterns that didn't work correctly with the TDFA implementation.

There are remaining problems when exceptions are thrown from JavaScript
callbacks.  See #483 for details on what's happening.  We need a better
answer for this in general.