Implement android builds and include assets
- Create `android-resources` directory that holds app assets
- Initialize `android-template` for using as a base for app builds
- Handler copies compiled `js` files into template and invokes `cordova`

A template directory is initialized with `cordova create` and assets
are copied in from `android-resources`. Includes main view HTML, CSS,
and App Icons.

The handler clones the template directory to `data/<mode>/android/` and
copies in the `js` files from the compilation step. `cordova build` is
invoked and the generated `.apk` file is fetched as a binary blob and
downloaded.