Move to a multi-file format for the guide.
At this point, each build-mode has a "shelf" file, which is a table
of different sections, each of which is its own markdown file.  The
contents now consists of a collapsible "accordion" control to browse
through multiple files.

Code could use some cleanup, e.g. to avoid refreshing the contents
and interrupting the animation when loading a new file.

Fixes #574.