JWT-based authentication using local SQLite3 database
This change implements an alternative authentication method that can be
configured to be used in place of the standard Google authentication
mechanism used by the official, live CodeWorld instance. It employs a
local SQLite3 accounts database with BCrypt password hashing. Stateless
authentication is established using signed JWT tokens generated on the
server and passed back by the browser.

An authenticated session is first established by the browser calling the
server's "/signIn" API using appropriate credentials which are validated
against the accounts database. The server then responds with a signed
JWT token in the "Authorization" header. The browser then subsequently
makes authenticated calls to the server by including the token in the
"Authorization" HTTP header. The server will respond by passing back the
API call result in the response body and a newly minted token in the
response's "Authorization" header. Refreshing the JWT token like this
results in a sliding expiry window. As currently implemented, these
token-based sessions expire after one hour of inactivity. If the session
expires, the user will be prompted to re-enter credentials.

This new authentication system is intended to be used in applications
where instructors cannot reasonably expect all students to have a Google
account. Note that local authentication enforces a very basic
5-character password policy.

Note that Google API's platform.js is loaded on-demand only if Google
authentication is configured. Also, we take a dependency on SweetAlert2
(in addition to SweetAlert) which supports multi-input UI more readily
than SweetAlert but which is not API-compatible with it.