JWT-based authentication using local SQLite3 database
This change implements an alternative authentication method that can be
configured to be used in place of the standard Google authentication
mechanism used by the official, live CodeWorld instance. It employs a
local SQLite3 accounts database with BCrypt password hashing. Stateless
authentication is established using signed JWT tokens generated on the
server and passed back by the browser.

An authenticated session is first established by the browser calling the
server's "/signIn" API using appropriate credentials using HTTP basic
authentication. These credentials are then validated against the
accounts database. The server then responds with a response body
containing two signed JWT tokens: an access token and refresh token. The
browser then makes subsequent authenticated calls to the server by
including the access token in the "Authorization" HTTP header (of
"Bearer" type). Assuming that the token has not yet expired, the server
will respond by passing back the API call result in the response body.
If the token has expired, the server will reject the request and the
client will make a call to the "/refreshToken" API along with the
refresh token. This will return a new access token and refresh token. If
the refresh token has also expired, the server will reject this and the
user will be prompted to re-enter credentials in the browser. It is the
browser's responsibility to sign out a session in order to invalid the
current refresh token (which is achieved using a monotonically
increasing token identifier).

This new authentication system is intended to be used in applications
where instructors cannot reasonably expect all students to have a Google
account. Note that local authentication enforces a very basic
5-character password policy.

Note that Google API's platform.js is loaded on-demand only if Google
authentication is configured. Also, we take a dependency on SweetAlert2
(in addition to SweetAlert) which supports multi-input UI more readily
than SweetAlert but which is not API-compatible with it.