Fix wonky dragging of resize bar.
This requires disabling pointer events on the iframe when a resize drag is
in progress.  Unfortunately, JQueryUI's layout doesn't let you listen for
that event, so instead, we are forced to settle for listening to DOM
changes and then looking for the existence of a cloned resize bar.  Yuck.

Fixes #691