Rescale the zoom slider to vary from 1/8 to 8x.
To line up with clicks, this required the click zoome amount to
become something weird: the tenth root of 8, which is about 1.23.
But that's still nicer than what was there before.

Even better would be to define some logical snap points (say, 12.5%,
25%, 33%, 50%, 75%, 100%, 125%, 150%, 200%, 400%, 800%), and have
the buttons proceed to the next snap point.  But that sounds
annoying to do, so I didn't do it.