Fix scaling of HistorySlider debug control
Previously, the slider was scaled as if there was a frame 0, which
actually doesn't exist.  So it was never actually possible to drag
the slider all the way left.  This is now fixed.

As a special-case, the slider sits all the way to the right when
there is only one frame, since its correct position is undefined.