More work on the guide.
As a side-effect of this, I fixed CodeWorld programs to not fail
when they are loaded in an iframe that is initially zero pixels in
size.  This is important because I want to embed CodeWorld examples
into iframes in collapsible notes within the guide, and collapsible
notes have zero size.