Run same click code in Inspect regardless of source.
In the old code, highlighting was added when building the tree (on a
click on the canvas), but separate code was run when a node was clicked
in the tree.  Some code was added on click to highlight the source, but
it was never run when clicking the canvas.

This change uses the same implementation regardless of which of the
representations is clicked.  That means the same source highlighting
happens in either case.

Fixes #730