Change behavior of `colored` on already-translucent pictures.
See #670 for details.  The current behavior has a surprising
discontinuity at RGBA(_, _, _, 0), which happens because any
non-transparent point (even if it's ALMOST completely transparent)
was just changed wholesale to the new color.  So:

    colored(colored(x, RGBA(0, 0, 0, 0)), blue) == blank
    colored(colored(x, RGBA(0, 0, 0, 0.001)), blue) == colored(x, blue)

despite the very small difference in values.  In fact, float
inaccuracies can cause small non-zero values, which now lead to
horribly wrong behavior.

The new behavior is that `colored` multiplies the alpha channel
of its first argument by the alpha of its color.  This is
consistent with a "fuzzy set" interpretation of the alpha
channel, where the alpha value records the extent to which a
point really belongs to the picture.

Fixes #670.