Add two new requirement rules.
notDefined verifies that there is no top-level definition for the
given name.

usesAllParams verifies that any named parameters defined for a function
are used in the body of the function.

These are still part of the stop-gap experimental language, and
shouldn't be trusted to remain in their current form.  I just need
them to continue the trial use.