Indeterminate result for requirements on unpareseable code.
This is now a corner case, and requires a buf in haskell-src-exts to
trigger, because requirements are only checked if GHC can build the
program.  Still, it's better to report that we can't do the eval,
rather than reporting success when we just don't know the answer.

Tested with https://code.world/#PEZDconO8BHxL6Ec3g_WDBg which does
exploit a haskell-src-exts bug.