Add wildcard names, and remove old wildcard functionality.
This removes the old wildcards, which have been replaced with
splice versions.  At the same time, it creates a new rule that any
name beginning with "__" will match any other name.  This is needed
to write a pattern that can match any function declaration, for
instance, since the function name isn't part of wither a pattern or
an expression, so it can't match a special form.