Only change velocity when bouncing, not position
This avoids violation of conservation of energy. Before, if the ball fell
below the border during a time step, it would have transformed potential
energy to kinetic energy during the fall, and then when it was teleported
back above the border by fence, it got the potential energy back and also
kept the kinetic energy.