Mark obsolete code based on editor generation, not hashed code.
This is less accurate, but it can be done instantly between
keystrokes, and that makes the UI more responsive.  Unlike code
parsing, it does make sense to detect and react to this right
away (there's no problem graying out the window when code is
incomplete).  So this is worth it.