Fix #856
This uses two heuristics to decide if a deferred error happens at
startup (when it should be treated as a compile error) or later
(when it would be shown as a separate message).  First, if the
program has any observable effect such as showing the canvas or
printing output, then it's a later effect.  Second, if it's been
longer than a second since startup.