Progress on #493.  This adds a MonadIO instance to CanvasM, which allows much more code to be shared between the various code paths.
My goal here is to help with creating a unit test for #681, which
are currently blocked because the driver code is not easily
testable.  With a bit more progress here, we can stub out the
browser-specific canvas bits (or even mock them, using monad-mock)
and test the surrounding driver logic.