Progress on #681
This doesn't fix the problem, but it does fix the specific causes
in the Wrapped type.

I don't even really understand why this works.  There's some real
oddness going on here with strictness.  Making all of the
non-wrapped fields strict fixes the bug, but it should not be
required.  I wish things made more sense.