Draw coordinate plan using alpha instead of thin lines.
Once upon a time, this caused problems when coloring the coordinate
plane, since that led to alpha being lost.  However, this was fixed
when the semantics of `colored` was changed to multiply by the old
alpha instead of overwriting it.  Now it's fine to use transparency
in a picture that will later be colored.

My hope is that this fixes #417, which I think came from aliasing
effects with lines less than one pixel wide.