Add screen size access to CanvasM monad.  Because ghcjs-base does not implement the CanvasRenderingContext2d.canvas property, I decided to plumb the size through the CanvasM monad for the JS implementation.  That could be changed.
This currently doesn't change anything.  However, it does pave the
road to make use of this information when drawing.  One use case is
an infinite coordinate plane a la #944, which would need to use the
screen size and current transformation to decide which subset of
the infinite plane to draw.