Split CodeWorld.Driver into CodeWorld.Driver and CodeWorld.EntryPoints
CodeWorld.EntryPoints contains the top-level entry points (drawingOf,
animationOf, etc.), and the implementation inside of CodeWorld primitives
of on-screen controls to manipulate these programs.  It does not contain
the core event loop or drawing code, which remain in CodeWorld.Driver.
The interface between these modules consists of the `runInspect` and
`runGame` functions.

This is one major step toward refactoring CodeWorld.Driver to reduce
the size of the module.  See #1009