Update the Reflex interface to separate UI events.
The idea is that in the future, we can change the event loop to
only listen for and react to events if they are actually used.
Previously, threading everything through CodeWorld's Event type
broke this feature of Reflex.

It also helps users to avoid exporting the CodeWorld Event type,
since there is a Reflex type of the same name.

See #1034