Fix cyclic dependency in reflex-with-inspect.
The problem was that debugState depended on userPicture, which depended
on logicalInput, which depended on debugState again.  The unnecessary
dependency was from debugState to userPicture, and it came from creating
debugState in connectInspect, which was doing too much.

In the new code, the debugState event is created before the main cycle,
and connected after the main cycle.  This means that the actual
instantiation of debugState happens outside connectInspect, and
connectInspect is merely passed an action to update the state.