XSS fix.
This converts everything (except the canvas recorder) to avoid cross-frame
access.  This means user code cannot access variables in the containing
frame, and that means the main XSS risk is resolved.  There is still the
fact that user code can access code.world cookies when run outside of the
iframe, but that's actually okay.  There aren't any high-risk code.world
cookies at all.  The main risk was the login info that was available from
member variables in the containing frameset.

Fixes #1193