Infrastructure for #1195
This isn't tested at all, but the idea is there: compileSource takes
a "module finder" parameter of type String -> IO (Maybe FilePath) that
finds source code for imports in user code.  If this returns non-Nothing
then the path it returns will be copied into the source directory,
modified so that it declares a module with the proper name, and then
added to the compiler command line.