Fix error messages from hash imports.
The first part of the change fixes parsing of error locations to
include locations in files other than program.hs.  Otherwise,
errors in other files are unparseable.

The second change is to fail when the imported file has diagnostics
that begin with "error:"  This includes actual errors, and rewritten
deferred errors, as well.  This is needed because GHC errors are all
treated as warnings in the wrapper, since it's GHC's exit code that
fails the build rather than any one message parsed from its output.
Unfortunately, this caused all errors from imported modules to be
swallowed because they seemed to be only warnings.

See #1236