Begin infrastructure for #1245
The change to codeworld-compiler is simple: just always try to
resolve Prelude with the import finder.  The greater part of the
commit is in two related changes:

1. To make testing possible with a Prelude replacement, I changed
the test harness so that it looks up any module in the current dir.
This required small adjustments to existing test cases.

2. To make it more convenient to create a custom Prelude, I exported
a new module called Core from codeworld-base.  This module exports
the minimum set of things necessary to support the language syntax
with RebindableSyntax enabled.  The pattern, then, is that a Prelude
replacement should be written like this:

    module Prelude (module Core, ...) where
    import Core
    ...

This way, basic syntax remains functional.