Add reflection and expand/correct discussion of transformations in guide.
One notable change is that we now use the term "preimage" to refer to
the picture input to a transformation.  This is common enough in math
curricula that it's likely to be helpful, and it's useful to have new
words to talk about the preimage versus the result of a transformation.

I did not include `clipped`, and I think it should be saved for
teacher-facing or developer-facing docs.  See #1254

Fixes #1275