Disable electric indent when indent is manually modified.
This is a mitigation for users who want to override the default
indentation.  If the indent is changed away from the default for
a line, it will not be automatically indented until it is returned
to the default, and the user leaves the line and returns (because
otherwise we'd catch cases where the indent is *accidentally*
restored).

Fixes #1280