Switch error-checks to compiling non-Main module.
The idea here is that when doing error checks, we'll tell GHC that
the main module is some other module (called literally
"Some_Other_Module"), so that GHC won't require that main (or program
in CodeWorld dialect) is defined here.  This makes things a bit more
ergonomic when building libraries for import-by-hash, but more
importantly, it avoids highlighting the first token of the file for
a silly error that really isn't located anywhere.

Fixes #1322