Implementation of #1362.
As of this commit, it is codeworld-api that is responsible for
setting up the canvas and styling for apps using the CodeWorld
graphics library.  The canvas is shown in the development
environment whenever anything is added to the DOM, no matter
which client-side framework is used to populate the page.

There is still some cleanup work to do.  In particular, the
screen-recording and debugging features of codeworld-api still
rely on JavaScript that's loaded and run in run.html.  Since
these features require interaction with the surrounding
development environment, that will probably remain mostly true,
but at least the script references can be loaded on demand,
rather than baked into the page when the app doesn't even use
CodeWorld at all.