Remove `squareRoot` and `absoluteValue` from the codeworld-base Prelude. The standard names have always been `sqrt` and `abs`.  Given work on issue #1380, we prefer to use an autocomplete redirect rather than add redundant names in the API.
The ultimate motivation was that having `squareRoot` in the API doesn't
work so well with other redirects, such as `square` -> `rectangle`,
because `squa` autocompletes to `squareRoot` under the direct-hit rule.
By moving `squareRoot` to another redirect instead of a full symbol, it
gets demoted so it can coexist with the redirect for `square`.

I have skipped the usual deprecation cycle, because I checked, and there
are no working programs using these functions written this year.  That's
a low enough volume that I think the risk would be higher that someone
starts using them during the deprecation.

@Powell-v2 FYI