Fix splitting of compiler diagnostics.
Previously we would just split compiler diagnostics on blank lines.
This works for ordinary errors and warnings, but doesn't work when
there are other messages, such as if a user specifies -ddump-simpl
or something like that.

This doesn't make -ddump-simpl work, because there is another bug
that unattributed lines are dropped.  However, it's a first step.