Stop running migration from run.sh
The migration referred to here is years old, now, and it's unlikely
anyone has a server that uses the old format.  If needed, the binary is
still there and built; it just doesn't run every time.

The motivation here is that on the main production server, that migration
actually takes over a minute, and that's all downtime because it cannot
run when the server is accepting requests.