Include compiler stdout in diagnostics
Certain GHC flags print to stdout rather than stderr, and this was
getting lost.  Even worse, it was filling a buffer, causing GHC to
block and time out, so the build failed.  Now we read both stdout
and stderr at the same time, and show both in the build output.